Using `{{ super() }}` for extending a block (#4309)
* Using `{{ super() }}` for extending a block

Yesterday we were struggling with adding a Zendesk widget (a simple JS) to our docs, because overriding any existing template block would lead to a broken site. Finally, we found this function which saved us. I wanted to share this, because it's not a common knowledge and it might be helpful for others.

* suggested correction