Fixes to "staying on the same page functionality"
This is #7350 with the tests removed, as requested in
https://github.com/squidfunk/mkdocs-material/pull/7350#issuecomment-2362039638 .

I do urge you to reconsider and merge something like #7350 (including
the tests I wrote) instead. Without tests, I suspect this code will
break like its [previous version did](
https://github.com/squidfunk/mkdocs-material/commit/ceacfe5a2c898585d5df610412f08565cebbb9cd)
. Once the code does break, it might be difficult to fix without tests.

Even if other parts of the theme work without tests, and even if you aren't in love with
how I set up the infrastructure in #7350 or #7338, I think this logic is complex enough
to need tests to be maintainable. Again, I wouldn't be able to check this code's correctness
without the tests that aren't included in this PR. 

Of course, this is your project to maintain, so do what works best for you. I just hope
this bugfix helps.

---------------------

Fixes #7226

Additionally, this allows using the version switcher even if the website
is published at a different URL than the `site_url`. In particular,
the version switcher should now work locally when serving the site with
`mike serve`.