Support fallback from `regex` package to standard `re` module
Support using the standard Python library `re` module if the `regex`
package is not available.  This improves the compatibility of plugins
with environments where `regex` is not or cannot be installed,
in particular when using PyPy.

The `regex` package is heavily relying on implementation details
of CPython and has a significant technical debt, that is preventing
the author from making the code more portable.  At the same time,
the standard library `re` module provides most of the functionality
used by mkdocs-material.

The `regex` package continues being used if it is installed
in the environment, and the change should therefore remain backwards
compatible.  When it is not installed, the `info` and `search` plugins
will use the standard library `re` module, therefore it becomes possible
to use them without having `regex` installed.  However, the Chinese
search support via `jieba` requires `regex`, and therefore will be
disabled if the fallback is used — much like it is currently disabled
if `jieba` is not installed.

I could not find any formal declaration of `regex` or `jieba`
requirements.  However, I did update the documentation to hint
on the necessity of having both `jieba` and `regex` installed for
the Chinese search to work.  I have also updated the old blog post — if
that is undesirable, I can revert it.

Fixes #8031